Fix grep dependency on locale for DDoc tests
By default, grep attempts to detect whether an input file is binary,
and if so, in the case of a match it simply prints "Binary file foo
matches". This detection depends on the current locale, as specified
in the environment.

The ddoc12 test contains Unicode characters. When the locale in the
environment is set to en_US.UTF8, this test passes. However, if the
enviroment variables are unset, the test fails, as grep considers
the file as binary. This causes the diff in ddocAny-postscript.sh
to fail, as it will compare the expected result with the string
"Binary file test_results/compilable/ddoc12.html matches".

This test failure was discovered with Digger, which uses a pristine
environment during its builds.